Move variant-analysis-filter-sort to shared directory
