+++ extensions/ql-vscode/src/variant-analysis/shared/variant-analysis-filter-sort.ts
@@ -1,6 +1,6 @@
-import { Repository, RepositoryWithMetadata } from "./shared/repository";
-import { parseDate } from "../common/date";
-import { assertNever } from "../common/helpers-pure";
+import { Repository, RepositoryWithMetadata } from "./repository";
+import { parseDate } from "../../common/date";
+import { assertNever } from "../../common/helpers-pure";
 
 export enum FilterKey {
   All = "all",
