+++ extensions/ql-vscode/src/view/variant-analysis/VariantAnalysisSkippedRepositoriesTab.tsx
@@ -8,7 +8,7 @@ import {
   compareRepository,
   matchesFilter,
   RepositoriesFilterSortState,
-} from "../../variant-analysis/variant-analysis-filter-sort";
+} from "../../variant-analysis/shared/variant-analysis-filter-sort";
 
 export type VariantAnalysisSkippedRepositoriesTabProps = {
   alertTitle: string;
