+++ extensions/ql-vscode/src/view/variant-analysis/__tests__/VariantAnalysisOutcomePanels.spec.tsx
@@ -16,7 +16,7 @@ import {
   createMockScannedRepo,
   createMockScannedRepos,
 } from "../../../../test/factories/variant-analysis/shared/scanned-repositories";
-import { defaultFilterSortState } from "../../../variant-analysis/variant-analysis-filter-sort";
+import { defaultFilterSortState } from "../../../variant-analysis/shared/variant-analysis-filter-sort";
 
 describe(VariantAnalysisOutcomePanels.name, () => {
   const defaultVariantAnalysis = {
