+++ extensions/ql-vscode/test/unit-tests/variant-analysis-filter-sort.test.ts
@@ -6,7 +6,7 @@ import {
   FilterKey,
   matchesFilter,
   SortKey,
-} from "../../src/variant-analysis/variant-analysis-filter-sort";
+} from "../../src/variant-analysis/shared/variant-analysis-filter-sort";
 
 /** A filterSortState that matches everything */
 export const permissiveFilterSortState = {
